Don't cause exception if the config fails to build
While failing to build kernel config sets the correct buildlog and
return code, in the follow-up code we are trying to copy a nonexistent
config file. This causes an exception to be raised, hiding the original
build failure behind what is considered an infrastructure problem.

We can solve this by wrapping the copying into a try-except block, but
we also need to include the `getrelease()` call in it -- this function
calls config generation if it didn't succeed before and it would fail
again. By including it in the block, we make sure the exception raised
by the copying is caught and the `getrelease()` is not called.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>