Pass args to cmd_build(), not merge
Pass the command line arguments directly to `cmd_build` and update
the state file directly without passing around the entire `cfg`
object.

The goal is to treat the state file like a state file and not a
config file. We should be doing just-in-time reads from and writes
to the state file (treating it more like a database).

Signed-off-by: Major Hayden <major@redhat.com>